% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_tree_prediction.R
\name{single_tree_prediction}
\alias{single_tree_prediction}
\title{Prediction of Testing Samples for single tree}
\usage{
single_tree_prediction(Single_Model, X_test, Variable_number)
}
\arguments{
\item{Single_Model}{Random Forest or Multivariate Random Forest Model of a particular tree}

\item{X_test}{Testing samples of size Q x N, Q is the number of testing samples and N is the number of features (same order and
size used as training)}

\item{Variable_number}{Number of Output Features}
}
\value{
Prediction result of the Testing samples for a particular tree
}
\description{
Predicts the output responses of testing samples based on the input regression tree
}
\details{
A regression tree model contains splitting criteria for all the splits in the tree and output responses of training 
samples in the leaf nodes. A testing sample using these criteria will reach a leaf node and the average of the 
Output response vectors in the leaf node is considered as the prediction of the testing sample.
}

